Blog Removal for Now!!
I am removign the blog, due to a now learning after much struggle and success in mental health and in life, that it has been a somatic experience and very intense. Much to figure out, but still coding, and curretly on tryhackme.com for cyber side defense. Thank you for reading this you're heros.